+++ components/amorphic/prepare-package.mjs
@@ -1,3 +1,4 @@
-import { createEsmModulePackageJson } from '@haventech/supertype/transformers/preparePackages.mjs';
+import { createCommonJsPackageJson, createEsmModulePackageJson } from '@haventech/supertype/transformers/preparePackages.mjs';
 
+createCommonJsPackageJson('./dist', 'cjs');
 createEsmModulePackageJson('./dist', 'esm');